Add files via upload
Four datasets that have nearly the same summary statistics
Based on the summaries we created with our statistical measures, these datasets are identical, but sometimes just looking at the summarized data can be misleading. Visualization can help us understand if the data is identical or not
The datasauRus creates plots with the Anscombe data in different shapes.